// src/store/my-account/action-types.ts
// Action enums
export enum ActionTypes {
  GET_ALL = 'GET_ALL',
  GET_DETAILS = 'GET_DETAILS',
  GET_MEMBERS = 'GET_MEMBERS',
  UPDATE_USER_DETAILS = 'UPDATE_USER_DETAILS',
  UPDATE_PASSWORD = 'UPDATE_PASSWORD',
  UPDATE_BILLING_ADDRESS = 'UPDATE_BILLING_ADDRESS',
  UPDATE_MEMBER_PASSWORD = 'UPDATE_MEMBER_PASSWORD',
  GET_ACCOUNT_DETAILS = 'GET_ACCOUNT_DETAILS',
  GET_IMPERSONATE_USER = 'GET_IMPERSONATE_USER',
}